app/requirements.txt
Flask==2.2.5
authlib==0.15.5
requests_oauthlib==1.3.0
python-dotenv==1.0.1
requests==2.32.3
